test(installer): split express prompt coverage and disclose vLLM container
