+++ ci/test-file-size-budget.json
@@ -7,7 +7,7 @@
     "src/lib/onboard/preflight.test.ts": 1905,
     "test/channels-add-preset.test.ts": 1871,
     "test/generate-openclaw-config.test.ts": 1984,
-    "test/install-preflight.test.ts": 4207,
+    "test/install-preflight.test.ts": 4006,
     "test/nemoclaw-start.test.ts": 5043,
     "test/onboard-messaging.test.ts": 2062,
     "test/onboard-selection.test.ts": 6888,
