+++ test/e2e-scenario/support-tests/e2e-live-project-config.test.ts
@@ -27,6 +27,7 @@ interface RootConfig {
 }
 
 const INSTALLER_INTEGRATION_TESTS = [
+  "test/install-express-prompt.test.ts",
   "test/install-preflight.test.ts",
   "test/install-preflight-docker-bootstrap.test.ts",
   "test/install-openshell-version-check.test.ts",
